add knowledge user and admin control

+++ src/knowledgeManage/action/deleteKnowledgeTreeNodeContentAction.java
@@ -35,12 +35,25 @@ public class deleteKnowledgeTreeNodeContentAction extends ActionSupport {
 
             //获取节点id
             String kid = request.getParameter("kid");
+            String userid = (String)session.getAttribute("userid");
 
             KnowledgeTreeDAO dao = new KnowledgeTreeDAOImpl();
-            Boolean rt = dao.deleteKnowledgeNode(kid);
 
+            //判断是否为作者用户
+            Boolean isValid = dao.isUser(kid, userid);
             JSONObject jsonObject = new JSONObject();
-            jsonObject.put("boolvalue", rt);
+
+            if(isValid){
+
+                Boolean rt = dao.deleteKnowledgeNode(kid);
+
+                if(rt)
+                    jsonObject.put("提示", "删除成功！");
+                else
+                    jsonObject.put("提示", "删除失败，请重试！");
+            }
+            else
+                jsonObject.put("提示", "删除失败，没有删除权限！");
 
             result = JSON.toJSONString(jsonObject);
